move settings component to settings
