+++ app/components/Views/Settings/GeneralSettings/index.test.js
